Draft of parser
This contains updates to the tokenizer which appeared when trying to
parse the W3C EBNF form.

This includes a slightly modified form of the EBNF as supplied by the
W3C.  The EXPONENT rule is expressed differently.  There are still
unknowns with respect to casing, as these are explained in statements
above the EBNF.

The parser parses the EBNF, but expresses rules in an inverse
direction.  This will be changed in a future commit.