Follow definition order in parsed EBNF
The parsed EBNF now follows the order in which the definitions were
made, rather than the inverse order.  This requires some
post-processing on the parsed content, hence it may be that our
interface changes in the near future.

When building the EBNF interpretation, it is easier to work with the
inverted form as it allows us to push items onto lists.  When working
with the resulting output, it is handier to have the order of the
definitions availaible.  This helps in interpreting the parsed rules,
as well as in interpreting the content.

It may be that we further extend our parsing to allow for simplified
forms.  This would mostly consist of removing spurious :paren_group
items.