Last argument in interpreter is for extra info
This last argument used to describe the elements which had to sent
back upwards.  Turns out that was not the most useful thing to do.  In
this iteration we ignored these elements for their current purpose,
and then reused the argument to describe information which has to be
sent down.  That information currently consists of whether or not the
current element is a terminal to parse, or not.