Emit structured content for each symbol match or consuming match
Previously, matches for symbols were emitted.  This may be
insufficient for bi-directionally syncing between solutions.  At
least, when the full string describing the current element can not be
used.  Prime examlpe is `'SELECT' ( 'DISTINCT' | 'REDUCED' )?` which
would not yield any string output in children, but require one of
`SELECT`, `SELECT DISTINCT` or `SELECT REDUCED` to be emitted.  If we
know only the symbol matches, there is no way to match these strings.
By remembering the strings, we should be able to reconstruct an
equivalent SPARQL query in all cases, and an equivalent EBNF in the
vast majority of cases.