Regen implementation
The regen implementation is notably cleaner than the ebnf_parser
interpreter.  Lessons from the regen can likely be taken in order to
optimize the ebnf_parser's interpreters.  Although wording isn't
always consistent, and some dead elements exist, other code should be
made consistent with this latest evolution.

We notice a lot of code duplication between the regen's processors,
and the InterpreterTerms.  There is most likely a common ground on
which these can be built with Elixir Behaviours.  We currently lack
the skill and know-how to make that happen.