More logical printing of SymbolMatch
The SymbolMatch printed with the 'symbol' property last.  This being
the most important property to scan for, we moved it way up front.
The default implementation made it tedious to interpret the output.
It is still somewhat tedious to interpret the output in the current
implementation, but it's way better still.