Custom rendering for Inspect of %Regen.Status{}
The EBNF syntax should be included in this object so it isn't
recalculated everywhere.  This has very interesting benifits in terms
of calculation, but it makes debugging near-impossible with 200k lines
output when parsing queries.  We're overriding the inspect protocol so
it doesn't include the syntax when the syntax is a large document.