Basic discovery of quads in INSERT DATA statements
The discovery is fairly defensive, but lacks some minor expressivenes
due to capping the terms on the terminal terms.  Turns out the
terminal terms aren't the cleanest separation in terms of information
flow.  Discovery does operate as expected, and yields a bunch of
quads.

A good starting-point is &Updates.quads_for_query/1