Added getting started.
Since the last time I contributed to this repository many things have changed
for the better. But the complexity has also grown so I have created this getting
started guide. It is mainly intended for myself and my fellow mu.semte.ch core
team fellows while we are developing this program.

Eventually this md file might grow out to be a fully fledge 'getting started'
page for potential non-core-team contributors.