Analyze INSERT WHERE queries and store through INSERT DATA
We can now analyze INSERT WHERE queries.  The WHERE portion is
executed separately, and is used to fill in the template of the INSERT
portion of the query.  This will allow us to identify changes
happening to the triplestore, as well as changing the target where the
content needs to be inserted.

This is not a fully-fledged version.  There are still quite a few
TODOs in the code.  However, the general way of working could likely
stay the same.