Fix retrieval of primitive values
The primitive values parsed correctly as per spec, but assumed the
contents should be yielded in a non-escaped form.  This is an
incorrect assumption.  The rest of the code assumes these elements can
be dropped right into a sparql query.  It also seems Virtuoso emits
"typed-literal", rather than "literal" for the "type" property.  Our
reading of the spec doesn't seem to allow for this, but it is now
parsed as if it were a typed literal value.