Implementation of Acl.Accessibility.ByQuery
This implementation required quite some changes to be made, as a
system may now potentially respond to multiple accessibility
constraints.  Furthermore, with the configuration given here and the
split between accessibility and resulting calculations, there is a
need to share the results of the accessibility to the
calculations (eg: we need to pass the current User ID to the
accessibility calculation).  This approach therefore puts in place the
building blocks for caching accessibilty.

The approach implemented here has some quirks.  Mostly TODOs which
need to be solved in a mechanical manner to ensure there are no holes
in the generated queries.