Perform access checks for sparql select queries
The select queries now also go by the access control lists.  This
ensures users can only see the contents which they are allowed to
see.  Some assumptions are made (like the graph in which the access
rights are stored), and there are some dangling issues (like
remembring which access rights a user has).  However, in the grand
scheme of things, this approach will let us move forward.