First draft of type discovery for update queries
Update queries have a set of unknown types.  We need to know the types
in order to insert the contents into the right graphs.  The current
implementation searches the triplestore to detect the types of unknown
resources.

There are a large set of upgrades needed in this code.  It is
notoriously slow at the moment, as it searches for far too much
content.  Because we know all the content that is inserted into the
triplestore, we could memoize the types in the triplestore and avoid
the round-trip.  Furthermore, we detect far too many types.  Much of
this is not necessary at all.  Lastly, there is an overly large set of
TODOs in the code which should be addressed in due time.