Send out, and consume, mu-authorization-groups header
This header contains the groups to which the current user has access.
We calculate them on the first run and share them together with the
result.  We expect the client to send this header to us on each
subsequent call, so specific optimizations can be made on this
syntax.  This content may be stored in the user's session (encryption
required when stored client-side), so the access rights can be held
over a longer period of time.  This allows us to calculate the access
rights once and reuse the results on a wide variety of calls.