Updates to forms in order to match the SPARQL1.1 test cases
The test cases assume that many spaces may occur between these words.
However, the EBNF writes them as a single sequence.  Being in an
airplane and not having a cached version of the proza above the EBNF,
I'm living under the assumption that the test cases are right.  As
such, the EBNF has been updated to allow for such cases.