Add helper functions which easen testing
The Parser now has a function which helps you check whether or not a
particular query can be fully parsed.  This system has the same
time-complexity as finding a solution, but it's nicer for testing
purposes.  We used this in an interactive IEx session in order to see
how far away we are from conforming with the SPARQL 1.1 spec.