Fixes for 'INSERT' 'DATA' and initial access right checking
The move from 'INSERT DATA' to 'INSERT' 'DATA' in the EBNF created a
bunch of issues.  This has been solved in the current commit.

Our goal of this commit was to ensure we would have a better way of
handling the access rights for users.  The new solution assumes you
can read/write to all graphs you have access to, and it matches the
WHERE clause of the INSERT DELETE WHERE with the graphs the current
user has access to.