Type cache and primitive version of clearing it
The type cache itself seems to operate as expected. Clearing can have
a myriad of race conditions beloning to it.  The clearing itself
should be done when the queries are executed.  The type of locking
necessary for this behaviour has not been figured out, and is
therefore left for the future.  Furthermore, the cache works
system-wide, it does not handle the cache based on the authorization
graphs of the current user (yet).

We doubt either of these will be an issue in simple applications as
the types are usually shared (shared uri's should denote the same
thing).  The race conditions should be minimal, as we don't expect the
types to change types once the types are defined.