Separation of authorization rules per useage
It is now required to specify for what use a particular access right
may be used.  There is a choice of :read, :write, :read_for_write.
With these access rights it becomes possible to have multiple rights,
but not to be allowed to read/write from/to all of them.