Add preliminary support for DELETE WHERE
Support for this clause is very similar to DELETE { clauses } WHERE {
clauses }.  Yet because the clauses in the delete portion are
differently structured than those in the where clause, there's some
strange conversion that needs to happen to make the EBNF correct.  We
do this by emitting and parsing the EBNF.

It should be verified that this is correct in all cases.  As in: we
should verify that the QuadPattern EBNF can only generate a subset of
GroupGraphPattern.  If there are other cases, then we need to tackle
them differently.  Should something be going wroing in this
implementation, then the emitted query will fail to execute, hence no
unauthorized data will be removed.