Don't require prefixes in query analyzer
The old solution required at least one prefix to be specified when
inserting data.  This is a non-existing requirement in the
specification, hence we're working around it now.  There may be other
places where this is assumed to be true.  In the future, we should
specify a set of default prefixes.  This would alleviate this issue.