Don't limit graph when calculating the authorizations.
It seems to be complex to configure the ByQuery if this holds.  We now
leave it open, knowing this will be slow in some many-graphs cases.
There will invariably be a need to configure this in the future,
possibly making transparent and smart decisions about it.  For now, we
drop the from_graph statement.