Interpretation of xsd prefixes regardless of options
It seems that many implementations assume the xsd prefix is specified
implicitly.  Altough we had assumed this was part of the SPARQL
syntax, we could not immediately find a reference to it.  Regardless,
it is an often used option, hence we're tackling it at the responsible
level at this time.