Cope with missing keys in options for Iri.from_prefix_string
This method assumed the options object would contain the necessary
items.  The code consuming this doesn't always supply these
arguments (eg: if no prefixes are given, the prefixes item is never
set).  Our new code copes with these situations as good as possible.
Note that not supplying information which is required will still
error.  For instance, when you don't have a default_graph, and want to
expand :hello, an error will still be thrown.