Log all the things (optionally)
There were a bunch of debug statements in the code.  Many of which
were not committed to a specific branch.  We have swapped these debug
statements out for the use of a Logger which can be configured.  The
current logging level is set to log some of the core elements (like
queries which are sent to the backend), but to ignore many
others (like the parsed queries).