Supply mu-auth-used-groups on SELECT queries
This is just a bogus implementation at the moment.  It assumes the
result could have come from all groups from which the current user had
access.  This is a fairly safe assumption, obviously, but it is most
likely overly broad.  This key will be kept in the future, but it's
implementation will become more meaningful.