Allow query to be sent in query params with POST query
Some clients may create a POST request, but use the query params of
the URL to store the query to be executed.  We didn't parse the query
in that case.  This case is now also supported.  Note that we don't
know whether the queries will match or not and that the resulting
queries may therefore execute slower.