Implementation of ResourceFormat matching queries by uri prefix
The Acl.GraphSpec.Constraint.ResourceFormat is a constraint like a
Acl.Constraint.Resource, but which doesn't match on the type, but
rather on the format of the URI.  This can be handy for internally
generated URIs.  These URIs sometimes don't receive a class in the
database, but can be used like general identifiers which are thrown
around.  Session identifiers are a good example of this.  With this
extention, mu-authorization can cope with this use-case.

Note that this module still contains a lot of duplication with the
Resource constraint.  It should be decoupled in a structured way.