Allow creating and filling diff templates
It is now possible to create a template_tree which contains the full
template, but with replaced variables.  It is also possible to convert
this tree into an array, which is much easier to fill in.  This array
can be matched with a query string, resulting in a set of variables
which should be replaced in the query.  Lastly, it is possible to
fill the template based on the discovered variables.