Generate string primitives as triple-double-quoted stings
When generating the EBNF for string primitives, we used to generate
double-quoted strings.  The issue with this is that some characters
which are allowed in triple-double-quoted strings (""") are not
allowed in strings with single-double-quotes (").  As the
triple quotes is a superset of the single quotes, we now revert to
rendering those.

The TODOs in these commits indicate work that should be done on order
te clean this up.  We don't expect to see new bugs arrive from this
code change for queries sent to the server.  A more complete solution
requires us to see where else these primitive values are being used.