Send delta messages to configured endpoints
We can now send delta messages to interested entities.  The
implementation occurs in the Delta module.  This code aims to be
cleaner Elixir code.  It has received type checking and other
documentation to see how we could benefit from this.

There are some missing elements in this approach.
- Parties which are interested in receiving the delta's are now informed
  immediately.  A sufficient mechanism for synchronization may need to
  be implemented before we use this to push changes to end-users.
- Parties which are interested in receiving the delta's are now
  informed for every change without knowing the initiator of the
  change.  In an ideal scenario, the involved parties would either
  know the initiator and/or only be informed about changes which they
  didn't cause themselves.
- Interested clients now receive all changes as they come in.  They
  may be overloaded or unable to handle the request.  Ideally speaking
  we have a single worker thread for each intersted party, and keep
  the changes stored for some time if the interested party is unable
  to respond to our requests.

All-in-all this first step is a good step forward.  It can be used for
real-world applications, but advancements may -and should- occur.