Allow external information on parsed query
The parsed queries contain the mapping between the query contents and
the formal SPARQL 1.1 syntax.  Some processing may want to attach
extra information to the syntax.  In order to do this, they would need
to wrap the original elements in new elements.  This, however, would
make it ill-suited for sharing contents between other components that
wish to manipulate the syntax.  The new format allows us to attach
extra information to the syntax.

Adding syntax works by creating a new variable-space, managed by maps
inside of the formal syntax (optimizations may be possible by using
linked lists when few variables or variable spaces exist).  This
variable-space can be queried in a similar way to maps.  The approach
is somewhat cleaner in lisp-like languages, and we migth be able to
make this clearer by using macro's in the future.  For now, this
approach seems sufficient.