Allow short-circuiting in map_matches
Walking over a query in order to rewrite it may have many reasons.
With the new support it is possible to :skip a step (which does not
traverse the current element further down, but continues processing),
as well as to :exit from the processing (this allows you to return a
value from deep down).  These steps allows us to apply our tree walker
to a broader spectrum of tasks.