Cleanup for build-time warnings, minor issues and logging
Some logging was left over in the GraphReasoner which is not necessary
anymore.  We're removing this logging for now so our state diff is
somewhat cleaner.  Also there was a potential bug in the
fully_processed? keyword, in which it seems the full graph was not
verified correctly.  This has been changed along the way.