Introduce QueryInfo struct in GraphReasoner
The introduction of this struct should help us share more data about
the query, with the query itself.  It is shared as a separate struct.
It is the intentation that extra information (like the semantic model,
or derived types) could be requested through this module, or related
modules, rather than through code embedded in the GraphReasoner.