Use config files in config/*.exs SPARQL_PORT env var for configuration
You can now set the SPARQL_PORT environment variable to set the public
SPARQL endpoint through an environment variable.  There are also the
config/dev.exs and config/prod.exs configuration files which allow to
set sensible defaults for development.

We should expect more of this type of configuration to be set through
this sort of configuration.