Allow variables in predicate portion for INSERT WHERE
When inserting data, the predicates themselves might consist of a
variable.  This should then be fetched as other content and filled in
the quad.

Looking at the code, the quads are already constructed in a rather
generic manner.  Hence we don't immediately see issues with allowing
them to be variables.  This is seemingly the result of extending the
code for INSERT WHERE without allowing for some of the fresh
capabilities.