Alter Logging.EnvLog.log method signature
The method signature of Logging.EnvLog.log assumed the name of the
key would be the first argument.  This was done in order to be in line
with the expectation of Logging.EnvLog.inspect.  This was done for the
latter in order to ensure it was easy to put this inspect into a
pipe.  It makes less sense to put this log into a pipe, hence we're
altering the signature to provide the key of where we want to log
first as this seems to be a more obvious way of working.

We'll see if this holds up in practice.