Configuration of delta and flexible configuration of user_groups
The user groups had to be defined in /config/config.ex, this would
yield problems to configure the delta, as that should be configured in
a config.ex file too.  We've chosen to separate away from the naming
in Elixir and instead allow users to specify user_groups.ex and
delta.ex for the respective configuration.  The former config.ex is
still supported for the time being.