Add logging of access rights and their processing
Access rights can be important to see as an end-user.  It is now
possible to log the access rights for each request.  We've also added
a way to inspect the processing of the access rights, as that tends
to be the step that follows when access rights don't seem to do their
thing.