Turn GraphReasoner into a working state
The GraphReasoner will work now as it receives arguments which make
sense.  It does not yield the desired results in some cases.

The current code does not correctly receive the UserGroups to apply,
and their configuration.

We have not checked the current code to correctly place properties of
code in the right group.  It may be that all propreties of a type are
assumed to live in the same graph.