Resolves all dialyzer warnings and formats code
These warnings should have proven to be about code that doesn't make
sense.  We molded the code so the warings don't apprear any longer.
This means that bugs may still be in the code, but normally no new
bugs should have been introduced by these changes.  Because of the
wide spectrum of the changes it may still be that we misinterpreted
something.  Testing this version of SEAS will be important.