Implements initial version of type reasoner with tooling
The type reasoner will reason on the content found in the query in
order to derive types (and possibly other information) which can later
be used to place things in the correct graphs.

The logic for type reasoning needs to be extended and is far too
complex to be placed in a single file as a minor function somewhere.
The type reasoner should also be more flexible in that it would be
easier to extend it to more than just variables, allowing us to attach
information to other elements (like possible values).