Provide warning ignoring for erroneous warnings of dialyzer
There seems to be a bug in dialyzer where it hints too many functions
don't exist.  We can ignore these warnings in the output with this
addition.

We should track this and see if we need it in the future.  One of the
issues realted to each other is at
https://github.com/jeremyjh/dialyxir/issues/258