Make TypeReasoner derive type from predicates unless explicit types
When no expilicit types have been given, we can resort to other
methods to derive the type of a statement.  In this approach we
consume the types derived from the predicates.  We combine all the
types from the predicates but don't go farther than that (eg: if the
object portion would be a variable we could be smart about the types
of that variable, but we're currently not).