Allow GraphReasoner information on more than subject variables
Subject variables contain most information.  They are the most common
element to attach information to.  Objects, and even predicates, might
also contain important information.  In a move to attach more
information to the query elements we're starting by attaching
information to predicates and objects.  This is a start for attaching
more advanced information.