Allow typing on non-variable subjects
Typing was determined based on the type of subject variables.  It is
now also supported to type on something else than a subject variable.
A constant URI used in the subject can now be requested from the
correct graphs.

In testing this, we found that we incorrectly took the range of
predicates into account, rather than their domain.  This has been
altered.