PoC Compatibility layer with Virtuoso workaround
It is now possible to set the database_compatibility environment
variable to "Virtuoso" to get a rewritten SPARQL query that changes
INSERT DATA to INSERT WHERE.

We should make this logic more intelligent.  This is a first version
without fixed API or internals but with a somewhat sensible
construction.  Code should be documented.