Add better error logging of unparseable database responses
When we execute queries to the database, and the response could not be
parsed, we now emit the respones on the console.  This can help in
debugging and adds minimal overhead.

We use a raise here but comment that we should use reraise.  When we
upgrade the elixir version of this application, we should alter that
call.