Handle situation where no access groups match
In the common setup a user will always have some access rights.  At
least the clear graph should be present.  However, in cases where the
access rights are chopped up and content is requested on a
per-access-right basis, this assumption does not necessarily hold.
This new update selects content from the empty
http://mu.semte.ch/graphs/empty graph in this situation.  The idea for
selecting from an empty graph is that we stay within the realm of the
specific database backend, meaning we can apply all sorts of necessary
transformations in a consistent manner, be it with the unnecessary
overhead of selecting from an empty graph.