Supply mu-call-id to database queries where plausible
The mu-call-id is now pushed forward to the database.  This allows us
to see which queries were the result from a particular user
interaction.  Notable exception to this is the cached classes of
resources.  These queries seem to live deeper in the stack, making it
too hard for us to pass that information through.