Add new endpoint /running-queries to detect long-running queries
The new endpoint keeps a registry of all queries that have ran which
haven't completed yet.  When calling the endpoint it will render the
contents in the response (in Elixir's native format, so not pretty),
and it will render the contents in the output stream.  This could
prove sufficient for the specific calls mu-authorization is currently
executing.

A similar endpoint should be built to explaint the incoming queries.