Add endpoint /processing-queries to list ongoing incoming queries
When a query arrives at mu-authorization it is split up and
processed.  This processing may take time.  In order to estimate
if/when queries are coming in which don't nicely end, we have a
process to list the processing of queries.  This is similar to the
/running-queries endpoint (which lists the queries as they're running
on the backend SPARQL endpoint).