Add exponential backoff to failing SPARQL queries
When a SPARQL query fails in production, there's a fair chance the
failure is coming from a backend error.  This is an approach to let us
catch these errors and retry the call multiple times.

It's not trivial to reach these errors in a simple development setup.
This feature will be launched on a more complex stack to check what
other errors pop up.