Support for cancelling query execution on the database
If a query fails to send, and we've already decided not to send an
answer to the user, we don't need to continue trying to send the query
to the database.  Executing the query is not necessary any longer.

This code should be updated so queries which may need to be executed
even though the response itself is not desired any longer are fully
ran (eg: in order to complete a partial update).