Move queries to compat layer, supporting timeouts
The actual querying of the triplestore has been moved into the compat
layer.  This should allow more flexible useage of the database's
capabilities.

One example is the non-standardized way in which Virtuoso can be
informed about the time it may take to produce a query result.  Such
feature can be extremely handy in cases where the user expects an
answer within a certain time-frame.  Virtuoso can be told to consume a
sane amount of time when calculating the query result.  This feature
will be extended to better inform the sparql endpoint about available
time.

Future work includes passing the total amount of time a query may take
into the query language, thereby allowing us to wait an amount of time
dictated by the consuming microservice.