Update InfoEndpoint to use ids in QueryInfo
The QueryInfo object was treated as a full object in order to identify
the record which would need changing.  This leads to subtle issues
when the object is updated down the line.  The new approach compares
the instances based on an ID, allowing code down the line to update
the instance and still allowing the parent to trigger removal.