Add experimental mu-request-timeout header
It is possible to set the expected timeout for a request with the
mu-request-timeout header.  This header will help indicate throughout
the processing whether or not the processing should still continue.
It also sends a 503 if the request takes longer than the specified
amount of time.

The spec isn't finalized, we don't know if this is the correct
approach but we know there should be an escape hatch.  This is the
escape hatch.  Triplestores don't need to abide this constraint and
not all processing is taken into account, so this is mostly a ballpark
number that should be seen as an indicator for the allowed query
complexity.  Even with these caveats, the implementation may help
guarantee processing in some cases and still be mildly agressive in
other cases.