Ensure quads are processed from the application graph
Users may specify a graph when trying insert where queries.  The graph
the user selects for inserting their data into should currently not be
taken into account.

The processing model for quads may change over time.  At this stage we
only understand the concept of an application graph.  As such any data
that is to be inserted and for which access rights need to be
validated should start calculating from the application graph.