Add ebnf syntax visualizer functionality
The EBNF syntax contains some logic.  This is a visualization of the
EBNF syntax.  Note that this syntax has an additional common group for
query and update.  The effective implementation used in
mu-authorization also contains some more efficient cutoffs by using
regular expressions.  This tree output can be handy to introspect the
valid structures of a SPARQL query.