Adding observer_cli as a dependency
This allows us to check what is going on in the system at a high
level.  It helps debugging remote systems which are contained inside a
docker instance.  In such cases it seems hard to forward graphical
systems to our local host.  However, inspecting through the terminal
does work like a charm.