Optionally provide an error when data will not be written
A configuration of mu-authorization may make some data not written
into the triplestore.  That's a real issue, especially with errors in
the configuration.  This new feature allows you to error out in such
case.  Meaning you'll get a warning that want you wanted to do is not
going to happen.  The full change is (in the current implementation)
not written to the triplestore (neither deletes nor inserts).  This is
only the case when the query will logically not have any effect, if
triples are inserted which already exist, that's fine.