Render default JSON-LD context when no query is supplied
This context will provide some information on the SPARQL endpoint.  Part
of this was implemented earlier but the response would not be rendered
correctly.  Our state was saved and the new implementation should render
said contents.

Reason for this implementation is the focus for being compatible with
Communica which uses this auto-discovery mechanism.  It may be that the
effective payload is changed to better follow the expectations for the
Communica integration.