Allow GraphSpec to specify the a :usage statement
The usage specifies what the access right can be used for (:read,
:write, :read_for_write).  It can now not only be specified on the
GroupSpec, but also on the GraphSpec.  When set on the GraphSpec, it
overrides the setting on the GroupSpec.  This should allow for a
migration from one model to the other.

We've opted for the term "usage" rather than the misspelled "useage"
that is used a level higher.  We consider this to be a better long-term
strategy.