Fix for langstring deletion
Langstrings were not extracted correctly from the WHERE block results of
the SPARQL query that is used to construct the DELETE DATA block for
DELETE WHERE queries.

For conditional DELETE queries we calculate the resulting WHERE block
before we effectively make changes to the triplestore.  The results of
this query are interpreted and transformed so they can be considered
QUADs on which an ACL can be applied.

The detection of a langstring had an incorrect match pattern leading to
a fallthrough.  This fallthrough caused the system to assume the string
was a regular string rather than a langstring, leading to data that was
not being deleted.